+++ aiotdlib/api/functions/parse_markdown.py
@@ -15,7 +15,7 @@ class ParseMarkdown(BaseObject):
     """
     Parses Markdown entities in a human-friendly format, ignoring markup errors. Can be called synchronously
     
-    :param text: The text to parse. For example, "__italic__ ~~strikethrough~~ **bold** `code` ```pre``` __[italic__ text_url](telegram.org) __italic**bold italic__bold**"
+    :param text: The text to parse. For example, "__italic__ ~~strikethrough~~ ||spoiler|| **bold** `code` ```pre``` __[italic__ text_url](telegram.org) __italic**bold italic__bold**"
     :type text: :class:`FormattedText`
     
     """